resolve code climate issues
